feat(SearchBox): add default width and height to buttons.
Since svg allows a width and height attribute, we can use those to provide a default width and height. This has a lower specificity than css defined width or height, because it's an svg.

Especially the situation where people don't use the Algolia theme is improved by this.

before|after
---|---
|