Attempting to make the virtual widget docs clearer
When I initially read the docs about virtual widgets, I did not understand why the virtual menu was being used and it was confusing. I have added a couple of sentences that should make understanding the use of the `VirtualMenu` more clear, which should, in turn, make the understanding of virtual widgets easier to understand.

If anything seems to be unclear or can be better, please let me know!