Escape HTML in changelog
It seems the changelog generation tool you've used doesn't take
into account that HTML-looking code is good to be escaped when
converted to Markdown. This is something which could be looked into separately.

Also, I'd prefer to go with fixed-width inline code elements for things like that: `InstantSearch` and `<InstantSearch>`.