chore(theme): use jsDelivr by default and set alias
We recommended unpkg at the begining, it was the only one serving other
files than JS from npm.

Now jsDelivr supports it by default since a long time, and we can inform
jsDelivr whats the default file to serve (to avoid style.min.css).

This will only work on the new release of the lib.

Also I updated the versions of the theme in the docs, might be worth
automating it or you'll have to do it next time too.