Fix bug with createInstantSearchServer/connectRange
When reducing over the searchParameters in findResultsState, searchParameters.getSearchParameters was being called with the wrong "this".

This led to a bug in connectRange where this._currentRange was undefined when getSearchParameters was being called.

If a "min" or "max" prop is provided to connectRange, an if block is executed inside getCurrentRefinement that tries to access properties on the undefined this._currentRange, causing an error