feat(DynamicWidgets): add fallbackWidget
This was brought up by @Shipow as a possible use case: you want to have dynamic widgets, but don't know what the facets could be in advance.

The fallbackWidget prop caters for this use case, allowing you to special-case some widgets (like hierarchical for example), but use the same widget for the rest.

```jsx
<ExperimentalDynamicWidgets fallbackWidget={RefinementList}>
  <HierarchicalMenu attributes={['hierarchy', 'subHierarchy']} />
</ExperimentalDynamicWidgets>

<ExperimentalDynamicWidgets fallbackWidget={({attribute}) => <CustomWidget facet={attribute} />}>
  <HierarchicalMenu attributes={['hierarchy', 'subHierarchy']} />
</ExperimentalDynamicWidgets>
```