feat(dynamicWidgets): send facets * and maxValuesPerFacet by default
FX-622

<!--
  Thanks for submitting a pull request!
  Please provide enough information so that others can review your pull request.
-->

**Summary**

<!--
  Explain the **motivation** for making this change.
  What existing problem does the pull request solve?
  Are there any linked issues?
-->

The default additional search parameters of `{ facets: ['*'], maxValuesPerFacet: 20 }` prevents an extra network request when widgets mount/unmount, as both queries will be identical, and thus cached.

If a refinement is set, this still results in two network requests, as the refinements can't be applied before it's known if the widget should be mounted.

**Result**

<!--
  Demonstrate the code is solid.
  Example: The exact commands you ran and their output,
  screenshots / videos if the pull request changes UI.

  You will be able to test out these changes on the deploy
  preview (address will be commented by a bot):

  1. the documentation site (/)
  2. a widget playground (/stories)
-->

When dynamic widgets is used without an explicit configure widget with `facets: ['*']` and `maxValuesPerFacet: highestLimit` an extra network request happens on load, which has a negative impact on performance. With the new additional parameters default value, widgets mounting or unmounting no longer require a separate network request

Two new parameters are added to `connectDynamicWidgets` and `dynamicWidgets`:

```js
<DynamicWidgets
  // ... other parameters
  // overridden with the default value
  facets={['*']}
  maxValuesPerFacet={20}
/>
```

The default behaviour can also be avoided, and the previous default (non-matching parameters, which does two network requests) can be done like this:

```js
<DynamicWidgets // ... other parameters
  // overridden default to ensure two requests are sent,
  // and that not all facets are requested
  facets={[]}
/>
```