fix(hooks): avoid importing from instantsearch.js' root
The root import of InstantSearch in the CJS build contains everything, including the components and widgets. These are not useful, and thus shouldn't be imported.

closes https://github.com/algolia/react-instantsearch/issues/3426

as React Native, through Metro, imports the CJS build, it ends up trying to build the InstantSearch.js components, which it can't as they're preact.

On the individual level this could be solved by putting "module" in the resolverMainFields: https://facebook.github.io/metro/docs/configuration#resolvermainfields, but we can't suggest people to make a Metro change, as for example Expo hides the Metro config.