instantsearch-hooks-server: use literal imports
Next.js build process (using pnpm) ends up not seeing these import
statements cause the import statements didn't have string literals.

The result was that it didn't put these dependencies in the nested
`node_modules` directory and thus this whole package breaks. Notably
this only fails in the production build and not the dev build.

This change is equivalent, we're just reformatting the code so the
literal import statements can be found by whatever tool Next.js is
using.